[OT-233][FIX]: 최근 시청내역 타입 위치 변경
